[GOHW-5]: :white_check_mark: add unit test to checkout service layer

+++ checkout/internal/api/checkout_v1/delete_from_cart_test.go
@@ -58,7 +58,7 @@ func TestImplementation_DeleteFromCart(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			mockCartService.EXPECT().
-				DeleteFromCart(ctx, tt.req.GetUser(), tt.req.GetSku(), tt.req.GetCount()).
+				DeleteFromCart(ctx, tt.req.GetUser(), tt.req.GetSku(), uint16(tt.req.GetCount())).
 				Return(tt.mockServiceErr).
 				Times(1)
